test: PathIntegrationTest 추가
