+++ src/test/java/subway/SubwayApplicationTests.java
@@ -1,0 +1,13 @@
+package subway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SubwayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
